+++ samples/FeatureFlags.AspNetCore/IApplicationBuilderExtensions.cs
@@ -3,6 +3,7 @@ using RimDev.AspNetCore.FeatureFlags;
 
 namespace FeatureFlags.AspNetCore
 {
+    //TODO: Remove this file
     public static class IApplicationBuilderExtensions
     {
         private static bool providerInitialized;